+++ RELEASENOTES.md
@@ -3,6 +3,7 @@
 ## 4.0.0
 
 - Fixed the problem that load opcode will trigger Implement Limitation on TRICORE.
+- Fixed the misbehavior of wasm memcpy on x86 backend
 
 ## 3.1.0